// resources/js/components/loggedin/components/wallet/user_detail.tsx
export class  UserPayment {
    firstname: string='Aksam' ;
    lastname: string='Lwanga' ;
    address1: string ='Kayunga';
    address2: string='' ;
    state: string ='town';
    country: string ='Uganda';
    code: string='2192';
    city: string='Kayunga' ;
    allow:boolean=true;
    cardname: string='';
    cardno: number=0;
    date: Date = new Date();
    cvv: number=0;
    amount:number=0;
 

}